changed home page to be the first thing u see
