+++ gamified-learning-frontend/app/page.tsx
@@ -1,103 +1,85 @@
-import Image from "next/image";
+"use client";
+import {useRouter} from "next/navigation";
+import { useEffect, useState } from "react";
 
+interface UserData {
+  username: string;
+  level: number;
+  xp: number;
+}
+
+interface LeaderboardUser {
+  userId: number;
+  username: string;
+  xp: number;
+}
 export default function Home() {
-  return (
-    <div className="font-sans grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
-      <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="font-mono list-inside list-decimal text-sm/6 text-center sm:text-left">
-          <li className="mb-2 tracking-[-.01em]">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] font-mono font-semibold px-1 py-0.5 rounded">
-              app/page.tsx
-            </code>
-            .
-          </li>
-          <li className="tracking-[-.01em]">
-            Save and see your changes instantly.
-          </li>
-        </ol>
+  const [user, setUser] = useState<UserData | null>(null);
+  const [challenges, setChallenges] = useState([]);
+  const [leaderboard, setLeaderboard] = useState<LeaderboardUser[]>([]);
+  const [categories, setCategories] = useState<string[]>([]);
+  const router = useRouter();
 
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:w-auto"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 w-full sm:w-auto md:w-[158px]"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+  useEffect(() => {
+    // TODO: get user data not just user 1
+    fetch("http://localhost:5180/api/users/1") 
+      .then((res) => res.json())
+      .then(setUser);
+
+    fetch("http://localhost:5180/api/challenges")
+      .then((res) => res.json())
+      .then(setChallenges);
+
+    fetch("http://localhost:5180/api/users/ordered")
+      .then((res) => res.json())
+      .then(setLeaderboard);
+
+    fetch("http://localhost:5180/api/challenges/categories")
+        .then((res) => res.json())
+        .then(setCategories);
+  }, []);
+
+  if (!user) return <p>Loading...</p>;
+
+  return (
+    <div className="p-6 text-white space-y-8">
+      <div>
+        <div className="flex justify-between items-center w-full">
+          <h1 className="p-4 text-3xl font-bold">Welcome back, {user.username} </h1>
+          <h2 className="p-4 text-3xl font-semibold cursor-pointer" 
+              onClick={() => router.push(`/profile`)}
+          >Profile</h2>
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+        <p>Level {user.level} • {user.xp} XP</p>
+      </div>
+
+      <section>
+        <h2 className="text-xl font-semibold mb-3">Available Categories</h2>
+        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
+            {categories.map((cat) => (
+            <div
+                key={cat}
+                onClick={() => router.push(`/categories/${encodeURIComponent(cat)}`)}
+                className="p-6 bg-gray-800 hover:bg-gray-700 rounded-lg cursor-pointer transition"
+            >
+                <h3 className="text-lg font-bold capitalize">{cat}</h3>
+                <p className="text-gray-400 text-sm">Explore challenges in {cat}</p>
+            </div>
+            ))}
+        </div>
+        </section>
+
+      <section>
+        <h2 className="text-xl font-semibold mb-3">Leaderboard</h2>
+        <ul>
+          {leaderboard.map((player, i) => (
+            <li key={player.userId} className="flex justify-between border-b border-gray-700 py-2">
+              <span>{i + 1}. {player.username}</span>
+              <span>{player.xp} XP</span>
+            </li>
+          ))}
+        </ul>
+      </section>
     </div>
   );
 }
